merge: adding contact feature to main #79 (#85)
* feat: create reviews app and add basic structure

Create reviews app and basic structure. to view and add reviews. #68

* feat: create template to add a review

Created template to add a review to a product. #70

* feat: add forms to handle the review collection

Adding form to collect review collection. #71

* feat: complete view to add review and urls

Completed view to add review and supporting urls. #71

* feat: template for add_review complete

Pending to further style the template but core functionality is
working fine. #70

* feat: complete view show all reviews by product

The will display all the reviews for the existing product displaying
the user name and the time the review was created. #72

* feat: display all users reviews in profile

Displaying all the user's reviews in profile, user will be able
to edit or delete reviews from that page. #73

* sytle: style my_reviews page to show up buttons

adding update and delete buttons to review.

* feat: add update review view

Adding update review page where a user will be able to update an
existing review. #74

* feat: add delete delete_review views

Adding delete review views to allow user to delete their own
reviews. #75

* feat: add contact app

Adding contact app and include it in settings and url #80

* feat: add forms.py to contact

This file will contain the contact form for the contact app. #83

* feat: add templates to contact app

Adding templates to hanle contact and contact succes pages. #84

* feat: add views to capture the contact form

Adding views to capture contact and to display contact success page #84

* feat: create contact form

adding forms.py to render the contact form #83

* feat: add urls.py to contact #80